Update param in findById function

+++ src/main/java/pl/vehiclerental/restapi/repository/UserRepository.java
@@ -12,7 +12,7 @@ import pl.vehiclerental.restapi.models.User;
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 
-    Optional<User> findById(int id);
+    Optional<User> findById(Long id);
 
     Boolean existsByUsername(String username);
